Moved code from Not_ify into MainWindow which has a fuller UI. Moved stacked windows to be the 3 panes, and the button to toggle as the settings button in bottom bar. Currently works by passing a reference when creating bottom bar, but could change to slots. Works fine like this anyway
